Refactor index.js to use Express Router
